Minor fixes to the combi-stick (#3184)
# About the pull request

Stops you from yanking the combi-stick if it is on your person
Unwields the combi-stick when you throw it

# Explain why it's good for the game

If you yanked the combi-stick on your person, it would duplicate it,
giving you two combi-sticks that are the same item.
The combi-stick wouldn't un-wield when you threw it, causing the
"offhand" object to stay in your hand even when thrown.

# Changelog

:cl:
fix: Fixes combi-stick duplicating and not un-wielding properly
/:cl: