Package Wrapper added to requisition back room (#3178)
# About the pull request

A player requested a package wrapper in req and it seems neat.

# Explain why it's good for the game

Neat stuff good. Wrapping things seems like a req task. If people do
stupid stuff with it we'll remove it.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: Morrow
maptweak: Package Wrapper added to requisition back room.
/:cl: