Alamo and Normandy now have a Reinforced Metal Ceiling (#3237)
Stops this kind of stuff from happening and weird tactics like
CAS/OB'ing the dropship because people know it can't be destroyed:


https://user-images.githubusercontent.com/24533979/236336813-8ecd8ba1-ba8e-4834-9acc-52814f466aec.mp4





The whole thing is indestructible might as well make the ceiling
indestructible as well.


![image](https://user-images.githubusercontent.com/24533979/236337629-32dc934f-d0c5-4f92-8885-2f45cd748236.png)

![image](https://user-images.githubusercontent.com/24533979/236337707-f047543b-0294-4301-b496-afcf90e6398a.png)


:cl: Hopek
balance: Alamo and Normandy now have a Reinforced Metal Ceiling meaning
that CAS and OB's can't penetrate it.
/:cl: