CLF Ship Nightmare Changes (#3199)
# About the pull request

updates the CLF ship's airlocks to new generic ones and adds one CLF
medic and one CLF engineer. (also a minor adjustment to the text about
sentries not having IFF (they do (the intro text doesnt even appear for
any survivors why has nobody re-added it))
(batu khan update coming soon)

# Explain why it's good for the game

The CLF ship should be very difficult to survive in, but not having
anyone that can make folding cades/barricades quickly means the best
option is to entirely ignore the cool new area of the map with potential
for a cool holdout/last stand, and instead immediately abandon it to
rush for a gap with the van. Having a medic also allows for the CLF
survivors to have a chance at being revived instead of one death
immediately putting them out of the game for good or meaning they'll
spend the rest of the round in the permabrig if they get revived by the
marines. I know a synth can spawn with them, but it's not a given at all
that a person is going to be rolling synthvivor every round that the
insert spawns in.

The doors were also outdated and took like 5 seconds to fully open,
annoyance for survivors, xenos and marines alike.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: IowaPotatoFarmer
maptweak: Switched the outdated CLF ship airlocks to new generic ones.
add: The CLF crashed ship insert on LV-624 now has one guaranteed medic
and one guaranteed engineer.
fix: The CLF crash intro text no longer includes inaccurate info about
the sentries.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>